Send chat button event to external.

+++ compile-sdk-mac.sh
@@ -34,12 +34,19 @@ xcodebuild -create-xcframework \
     -output ios/sdk/out/JitsiMeetSDK.xcframework
 rm -rf ios/sdk/out/WebRTC.xcframework
 cp -a node_modules/react-native-webrtc/apple/WebRTC.xcframework ios/sdk/out
-rm -rf ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/WebRTC.xcframework
-rm -rf ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/JitsiMeetSDK.xcframework
-cp -a node_modules/react-native-webrtc/apple/WebRTC.xcframework ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/
-cp -a ios/sdk/out/JitsiMeetSDK.xcframework ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/
 
-rm -rf ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/WebRTC.xcframework
-rm -rf ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/JitsiMeetSDK.xcframework
-cp -a node_modules/react-native-webrtc/apple/WebRTC.xcframework ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/
-cp -a ios/sdk/out/JitsiMeetSDK.xcframework ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/
+if [ -d ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks ]; then
+  echo "Copying to jitsimeet-app ..."
+  rm -rf ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/WebRTC.xcframework
+  rm -rf ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/JitsiMeetSDK.xcframework
+  cp -a node_modules/react-native-webrtc/apple/WebRTC.xcframework ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/
+  cp -a ios/sdk/out/JitsiMeetSDK.xcframework ../jitsimeet-app/ios/Pods/JitsiMeetSDK/Frameworks/
+fi
+
+if [ -d ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks ]; then
+  echo "Copying to CircleIt ..."
+  rm -rf ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/WebRTC.xcframework
+  rm -rf ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/JitsiMeetSDK.xcframework
+  cp -a node_modules/react-native-webrtc/apple/WebRTC.xcframework ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/
+  cp -a ios/sdk/out/JitsiMeetSDK.xcframework ../CircleIt/ios/Pods/JitsiMeetSDK/Frameworks/
+fi
